requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
folium>=0.14.0
streamlit-folium>=0.13.0
numpy>=1.24.0
geopy>=2.3.0
scikit-learn>=1.3.0
googlemaps>=4.10.0
polyline>=1.4.0
python-dotenv>=1.0.0